pyproject.toml
[tool.poetry]
name = "auctionbot"
version = "1.0.0"
description = "Discord bot that posts new bids on SuperRare"
authors = ["Ruud Schroën"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.0.0"
python-dotenv = "^0.19.1"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}
isort = "^5.9.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
